fedimint_dbtool

Module dump

Macros

push_db_pair_items_no_serde 🔒

Structs

DatabaseDump

Structure to hold the deserialized structs from the database. Also includes metadata on which sections of the database to read.

SerdeWrapper 🔒